Add Environment and Current for lightweight DI
Based on the Point Free talks on this https://www.pointfree.co/episodes/ep16-dependency-injection-made-easy

I wanted something lightweight given most of the code is a bunch of functions that take values, grouped together in some classes. Didn't want to add protocols and or use constructor injection or something. This seems like a decent way to be able to swap specific functionality out for testing. Will need to continue the pattern with FileManager in order to test XcodeInstaller, but this will allow testing the overall installation logic without having to have a XIP laying around.