Fetch prerelease Xcodes when updating
This involves fetching the developer.apple.com/downloads HTML and scraping it for the right links. SwiftSoup thankfully makes this really easy without requiring any system libraries. In addition, the Xcode initializer was simplified. An extension on Version adds the ability to parse and output the marketing format version used on /downloads.

Closes #6.