Only spin the RunLoop for async commands
The primary motivation for this was that Guaka doesn't provide a hook to `exit()` after printing the help message so it'll just hang. I explored waiting on the main thread for PMKFinalizers to finish when they weren't dispatched, but I'm not confident that it wouldn't deadlock somehow. This is simpler and fixes the original problem, so let's try it out for a bit.