Consolidate versions from both download sites
Starting with Xcode 11 beta 6, developer.apple.com/download and developer.apple.com/download/more both list some pre-release versions of Xcode.
Previously pre-release versions only appeared on developer.apple.com/download.
/download/more doesn't include build numbers, so we trust that if the version number and prerelease identifiers are the same that they're the same build.
If an Xcode version is listed on both sites then prefer the one on /download because the build metadata is used to compare against installed Xcodes.

Because pre-release versions that are installed also have build metadata, but this won't be the case for pre-release versions listed on /download/more, we also need to replace versions that are available with the installed ones with build metadata before printing them.