Improve command documentation
It was previously not clear how to specify versions of Xcode to install. The current version argument parsing is sort of incidentally a little fuzzy, and so there are sometimes different ways to specify a single pre-release version like "11.2.1 GM Seed", such as "11.2.1GM Seed". This change adds examples to the command help and to the project README in order to give examples of install commands for pre-release versions. I don't think this is totally sufficient and there is more to be done, but it's a first step.

https://github.com/RobotsAndPencils/xcodes/issues/80