Update Apple auth URLs
The URLs we were using were originally from the xcode-install project and had been working up until today (2019/11/12). These URLs were updated in that project (actually in fastlane, which it uses for auth) about 7 months ago but I didn't notice this. I started to look in a web inspector while signing in to developer.apple.com and saw that the Olympus session URL changed. Dave DeLong noted that a slightly different URL worked for the iTC service key, which is also what xcinfo uses.

https://github.com/fastlane/fastlane/pull/14509
https://github.com/RobotsAndPencils/xcodes/issues/83#issuecomment-553207943
https://github.com/xcodereleases/xcinfo/blob/e51f022db3584be884b5d10d8b161ac26dbbee0b/Sources/OlympUs/OlympUs.swift#L213