Remove LegibleError
While it seems valuable for errors without localized descriptions, it provides worse strings for NSErrors because it intentionally doesn't return their localized descriptions when they start with "The operation couldn’t be completed." Maybe this is due to the state of Foundation on Linux. In any case, I don't think this is providing as much value as I thought it would, so we'll revert to using the localized descriptions.

Fixes #79.