Added `--delete-xip` flag to `xcodes install`
After Pull Request #60, `xcodes` always moves Xcode's .xip to the Trash after installation. This is problematic for scripts, because on modern macOS versions, [Full Disk Access is required to programatically delete the Trash](https://apple.stackexchange.com/questions/376916/cannot-ls-trash-in-the-terminal-in-catalina-operation-not-permitted).

To solve this without reducing security, added a new `--delete-xip` flag that automatically deletes the .xip after a successful installation.

- This is done this way to preserve the current behaviour (moving the .xip to the Trash) by default, [as it was originally intended](https://github.com/RobotsAndPencils/xcodes/issues/56#issuecomment-512919385).

This closes #185.