Quick fix to ignore new iOS 18 runtime downloadables
With introduction if Xcode 16 and iOS 18 simulators, a breaking change
for xcodes was introduced through the simulators API. ("https://devimages-cdn.apple.com/downloads/xcode/simulators/index2.dvtdownloadableindex")

In order to keep current functionality working, I made 2
modifications:
1. Make runtime source optional. This is needed as iOS 18 downloadable
does not have that value.
2. Add support for new `cryptexDiskImage` content type. As iOS 18
downloadable comes with new type of `contentType`, we should parse it
in order to be able to process rest of the response.

In RuntimeInstaller, I've added logic to handle new optional and new
content type be throwing error if those downloadable were to be used.

Test cases and fixtures were also updated to handle these new cases.